Add BroadcastManager class that takes away JUCE dependent from GameState.h
