+++ CMakeLists.txt
@@ -72,11 +72,11 @@ juce_add_gui_app(SoundTests
 # Finally, we supply a list of source files that will be built into the target. This is a standard
 # CMake command.
 
-# add_library(Common OBJECT src/GameState.cpp)
+add_library(Common OBJECT src/GameState.cpp)
 
 target_sources(MusicalChess
     PRIVATE
-        src/GameState.cpp
+        src/BroadcastManager.h
         src/ChessboardGUI.h
         src/ChessboardGUI.cpp
         src/Main.cpp
@@ -95,7 +95,6 @@ target_sources(MusicalChess
 
 target_sources(Tests
     PRIVATE
-        src/GameState.cpp
         src/MainProcessor.cpp
         src/Wavetable.h
         src/SoundProcessor.cpp
@@ -106,12 +105,10 @@ target_sources(Tests
 
 target_sources(TerminalChess
     PRIVATE
-        src/GameState.cpp
         tests/TerminalChess.cpp)
 
 target_sources(SonifierTest
 	PRIVATE
-        src/GameState.cpp
 		src/SoundProcessor.cpp
         src/MainProcessor.cpp
         src/Scheduler.cpp
@@ -121,7 +118,6 @@ target_sources(SonifierTest
 
 target_sources(SoundTests
     PRIVATE
-        src/GameState.cpp
         src/SoundProcessor.cpp
         src/Scheduler.cpp
         src/MainProcessor.cpp
@@ -187,7 +183,7 @@ target_compile_definitions(SoundTests
 target_link_libraries(MusicalChess
     PRIVATE
         # GuiAppData            # If we'd created a binary data target, we'd link to it here
-        # Common
+        Common
         juce::juce_gui_extra
         juce::juce_audio_utils
     PUBLIC
@@ -197,7 +193,7 @@ target_link_libraries(MusicalChess
 
 target_link_libraries(Tests
     PRIVATE
-        # Common
+        Common
         juce::juce_gui_extra
         juce::juce_audio_utils
     PUBLIC
@@ -207,7 +203,7 @@ target_link_libraries(Tests
 
 target_link_libraries(TerminalChess
     PRIVATE
-        # Common
+        Common
         juce::juce_gui_extra
         juce::juce_audio_utils
     PUBLIC
@@ -218,7 +214,7 @@ target_link_libraries(TerminalChess
 
 target_link_libraries(SonifierTest
     PRIVATE
-        # Common
+        Common
         juce::juce_gui_extra
         juce::juce_audio_utils
     PUBLIC
@@ -228,7 +224,7 @@ target_link_libraries(SonifierTest
 
 target_link_libraries(SoundTests
     PRIVATE
-        # Common
+        Common
         juce::juce_gui_extra
         juce::juce_audio_utils
     PUBLIC
